fixed bug in owner collection

+++ app/Http/Controllers/OwnerCollectionController.php
@@ -188,7 +188,7 @@ class OwnerCollectionController extends Controller
       
          for($i=0; $i<=$counter; $i++)
          {
-            $collection = (int) $request->input("collection_amount_".$i);
+            $collection = (double) $request->input("collection_amount_".$i);
 
             $form = $request->form;
